refactor: dashboard, focus, services and repositories

+++ src/EasyFlow.Application/Tags/DeleteTag.cs
@@ -12,15 +12,24 @@ public sealed class DeleteTagCommand : IRequest<Result<bool>>
 public sealed class DeleteTagCommandHandler : IRequestHandler<DeleteTagCommand, Result<bool>>
 {
     private readonly ITagsRepository _tagsRepository;
+    private readonly IGeneralSettingsRepository _generalSettingsRepository;
 
-    public DeleteTagCommandHandler(ITagsRepository tagsRepository)
+    public DeleteTagCommandHandler(ITagsRepository tagsRepository, IGeneralSettingsRepository generalSettingsRepository)
     {
         _tagsRepository = tagsRepository;
+        _generalSettingsRepository = generalSettingsRepository;
     }
 
     public async Task<Result<bool>> Handle(DeleteTagCommand request, CancellationToken cancellationToken)
     {
         var tag = request.Tag!;
+
+        var allSettings = await _generalSettingsRepository.GetAsync();
+        var settings = allSettings[0];
+        if (settings.SelectedTagId == tag.Id)
+        {
+            return Result<bool>.Failure(TagsErrors.CanNotDeleteSelectedTag);
+        }
 
         var result = await _tagsRepository.DeleteAsync(tag);
         if (!result)
@@ -36,4 +45,7 @@ public static partial class TagsErrors
 {
     public static readonly Error DeleteFail = new($"Tag.DeleteFail",
        "Failed to delete the tag");
+
+    public static readonly Error CanNotDeleteSelectedTag = new($"Tag.CanNotDeleteSelectedTag",
+       "Failed to delete the tag. You can't delete the selected tag");
 }
